\name{circda-package}
\alias{circda-package}
\docType{package}
\title{
Circular Data Analysis
}

\description{
Functions to perform maximum likelihood estimation, model-based clustering,
discriminant and regression analysis with a circular response variable.
}

\details{
\tabular{ll}{
Package: \tab circda\cr
Type: \tab Package\cr
Version: \tab 1.1 \cr
Date: \tab 2026-09-14 \cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr}
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>.
}

\references{
Tsagris M. and Alenazi A. (2019). Comparison of discriminant analysis methods on the sphere.
Communications in Statistics: Case Studies, Data Analysis and Applications, 5(4): 467--491.

Tsagris M., Papastamoulis P. and Kato S. (2025). Directional data analysis using the spherical Cauchy and the Poisson kernel-based distribution.
Statistics and Computing, 35:51.
}





